+++ juniper-from-schema-code-gen/src/ast_pass/ast_data_pass.rs
@@ -1,20 +1,23 @@
-use crate::ast_pass::{schema_visitor::SchemaVisitor, type_name};
-use graphql_parser::schema::{Document, *};
-use std::collections::{HashMap, HashSet};
+use crate::ast_pass::{
+    directive_parsing::{DateTimeScalarType, ParseDirective},
+    error::{Error, ErrorKind},
+    schema_visitor::SchemaVisitor,
+    type_name, EmitError,
+};
+use graphql_parser::{
+    schema::{Document, *},
+    Pos,
+};
+use std::collections::{BTreeSet, HashMap, HashSet};
 
 pub struct AstData<'doc> {
     interface_implementors: HashMap<&'doc str, Vec<&'doc str>>,
-    special_scalars: HashSet<&'doc str>,
+    user_scalars: HashSet<&'doc str>,
     enum_variants: HashSet<&'doc str>,
     input_object_field_types: HashMap<&'doc str, HashMap<&'doc String, &'doc Type>>,
-}
-
-impl<'doc> From<&'doc Document> for AstData<'doc> {
-    fn from(doc: &'doc Document) -> Self {
-        let mut data = AstData::new();
-        data.visit_document(doc);
-        data
-    }
+    errors: BTreeSet<Error<'doc>>,
+    raw_schema: &'doc str,
+    include_time_zone_on_date_time_scalar: bool,
 }
 
 impl<'doc> SchemaVisitor<'doc> for AstData<'doc> {
@@ -28,8 +31,20 @@ impl<'doc> SchemaVisitor<'doc> for AstData<'doc> {
     }
 
     fn visit_scalar_type(&mut self, scalar: &'doc ScalarType) {
-        let name = &*scalar.name;
-        self.special_scalars.insert(name);
+        match &*scalar.name {
+            name @ "DateTime" => {
+                let args = self.parse_directives(DateTimeScalarType(scalar));
+                if args.with_time_zone {
+                    self.include_time_zone_on_date_time_scalar = true;
+                } else {
+                    self.include_time_zone_on_date_time_scalar = false;
+                }
+                self.user_scalars.insert(name);
+            }
+            name => {
+                self.user_scalars.insert(name);
+            }
+        };
     }
 
     fn visit_enum_type(&mut self, enum_type: &'doc EnumType) {
@@ -47,16 +62,33 @@ impl<'doc> SchemaVisitor<'doc> for AstData<'doc> {
 }
 
 impl<'doc> AstData<'doc> {
-    fn new() -> Self {
-        Self {
-            interface_implementors: HashMap::new(),
-            special_scalars: HashSet::new(),
-            enum_variants: HashSet::new(),
-            input_object_field_types: HashMap::new(),
+    pub fn new_from_schema_and_doc(
+        raw_schema: &'doc str,
+        doc: &'doc Document,
+    ) -> Result<Self, BTreeSet<Error<'doc>>> {
+        let mut data = Self::new(raw_schema);
+        data.visit_document(doc);
+
+        if data.errors.is_empty() {
+            Ok(data)
+        } else {
+            Err(data.errors)
         }
     }
 
-    pub fn get_interface_implementor(&self, name: &str) -> Option<&Vec<&str>> {
+    fn new(raw_schema: &'doc str) -> Self {
+        Self {
+            interface_implementors: Default::default(),
+            user_scalars: Default::default(),
+            enum_variants: Default::default(),
+            input_object_field_types: Default::default(),
+            errors: Default::default(),
+            raw_schema,
+            include_time_zone_on_date_time_scalar: true,
+        }
+    }
+
+    pub fn get_implementors_of_interface(&self, name: &str) -> Option<&Vec<&str>> {
         self.interface_implementors.get(name)
     }
 
@@ -64,8 +96,16 @@ impl<'doc> AstData<'doc> {
         self.is_scalar("Date")
     }
 
-    pub fn date_time_scalar_defined(&self) -> bool {
-        self.is_scalar("DateTime")
+    pub fn date_time_scalar_definition(&self) -> Option<DateTimeScalarDefinition> {
+        if self.is_scalar("DateTime") {
+            if self.include_time_zone_on_date_time_scalar {
+                Some(DateTimeScalarDefinition::WithTimeZone)
+            } else {
+                Some(DateTimeScalarDefinition::WithoutTimeZone)
+            }
+        } else {
+            None
+        }
     }
 
     pub fn uuid_scalar_defined(&self) -> bool {
@@ -77,7 +117,7 @@ impl<'doc> AstData<'doc> {
     }
 
     pub fn is_scalar(&self, name: &str) -> bool {
-        self.special_scalars.contains(name)
+        self.user_scalars.contains(name)
     }
 
     pub fn is_enum_variant(&self, name: &str) -> bool {
@@ -123,4 +163,20 @@ impl<'doc> AstData<'doc> {
         let type_ = field_map.get(field_name)?;
         Some(type_name(&type_))
     }
+}
+
+impl<'doc> EmitError<'doc> for AstData<'doc> {
+    fn emit_non_fatal_error(&mut self, pos: Pos, kind: ErrorKind<'doc>) {
+        let error = Error {
+            pos,
+            kind,
+            raw_schema: &self.raw_schema,
+        };
+        self.errors.insert(error);
+    }
+}
+
+pub enum DateTimeScalarDefinition {
+    WithTimeZone,
+    WithoutTimeZone,
 }