+++ juniper-from-schema-code-gen/src/ast_pass/error.rs
@@ -166,10 +166,32 @@ impl<'doc> fmt::Display for Juniper<'doc> {
 }
 
 #[derive(Debug, Clone, Eq, PartialEq, Ord, PartialOrd)]
+pub enum Scalar<'doc> {
+    WrongNumberOfArgs(usize),
+    InvalidKey(&'doc str),
+}
+
+impl<'doc> fmt::Display for Scalar<'doc> {
+    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
+        match self {
+            Self::WrongNumberOfArgs(count) => write!(
+                f,
+                "Wrong number of arguments. Expected `1`, got `{}`",
+                count
+            ),
+            Self::InvalidKey(name) => {
+                write!(f, "Invalid key `{}`. Expected `with_time_zone`", name)
+            }
+        }
+    }
+}
+
+#[derive(Debug, Clone, Eq, PartialEq, Ord, PartialOrd)]
 pub enum UnsupportedDirectiveKind<'doc> {
     Deprecation(Deprecation<'doc>),
     Ownership(Ownership<'doc>),
     Juniper(Juniper<'doc>),
+    Scalar(Scalar<'doc>),
     InvalidType {
         actual: ValueType,
         expected: ValueType,
@@ -182,6 +204,7 @@ impl<'doc> fmt::Display for UnsupportedDirectiveKind<'doc> {
             Self::Deprecation(inner) => write!(f, "{}", inner),
             Self::Ownership(inner) => write!(f, "{}", inner),
             Self::Juniper(inner) => write!(f, "{}", inner),
+            Self::Scalar(inner) => write!(f, "{}", inner),
             Self::InvalidType { expected, actual } => {
                 write!(f, "Invalid type. Expected `{}`, got `{}`", expected, actual)
             }